+++ molecularnodes/assets/data.py
