+++ molecularnodes/entities/trajectory/__init__.py
@@ -1,4 +1,4 @@
 from . import selections
 from .io import load, load_oxdna
 from .base import Trajectory
-from .dna import OXDNA
+from .oxdna import OXDNA
