--- molecularnodes/operators/__init__.py
@@ -1,3 +1,0 @@
-from . import node_add_buttons
-
-CLASSES = node_add_buttons.CLASSES
