+++ molecularnodes/utils.py
@@ -1,14 +1,12 @@
-import os
 import sys
 import numpy as np
 import json
 
 from pathlib import Path
-from math import floor
 from mathutils import Matrix
 
 ADDON_DIR = Path(__file__).resolve().parent
-MN_DATA_FILE = os.path.join(ADDON_DIR, "assets", "MN_data_file_4.2.blend")
+MN_DATA_FILE = ADDON_DIR / "assets/MN_data_file_4.2.blend"
 
 
 def add_current_module_to_path():
